### 2024-11-13 #7107 @cadojo
> 👋 @cadojo - Would you be willing to review this submission to JOSS?

I can also review thes paper, @crvernon!
