### 2024-11-13 #7473 @editorialbot
[REVIEW]: robot_collision_checking: A Lightweight ROS 2 Interface to FCL (Flexible Collision Library)
**Submitting author:** @mazrk7 (<a href="http://orcid.org/0000-0002-7672-940X">Mark Zolotas</a>)
**Repository:** https://github.com/philip-long/robot_collision_checking
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
