### 2024-11-24 philip-long/robot_collision_checking#4 @cadojo
error: expected type-specifier fcl
After installing `licccd` and `fcl` from source, I am unable to build `fcl_interface` with `colcon`. I get the error shown below. For reference, a demo which shows this example with the `ros:humble` container is available at [cadojo/robot_collision_checking_demo](https://github.com/cadojo/robot_collision_checking_demo).

```
robot_collision_checking on  main
⬢ [Docker] ❯ colcon build
Starting >>> robot_collision_checking
--- stderr: robot_collision_checking
/workspaces/robot_collision_checking/src/robot_collision_checking/src/fcl_interface.cpp: In function ‘robot_collision_checking::FCLCollisionGeometryPtr robot_collision_checking::fcl_interface::createCollisionGeometry(const Octomap&)’:
/workspaces/robot_collision_checking/src/robot_collision_checking/src/fcl_interface.cpp:111:35: error: expected primary-expression before ‘(’ token
  111 |     return FCLCollisionGeometryPtr(new fcl::OcTreed(tree));
      |                                   ^
/workspaces/robot_collision_checking/src/robot_collision_checking/src/fcl_interface.cpp:111:40: error: expected type-specifier
  111 |     return FCLCollisionGeometryPtr(new fcl::OcTreed(tree));
      |                                        ^~~
gmake[2]: *** [CMakeFiles/fcl_interface.dir/build.make:76: CMakeFiles/fcl_interface.dir/src/fcl_interface.cpp.o] Error 1
gmake[1]: *** [CMakeFiles/Makefile2:159: CMakeFiles/fcl_interface.dir/all] Error 2
gmake: *** [Makefile:146: all] Error 2
---
Failed   <<< robot_collision_checking [3.27s, exited with code 2]

Summary: 0 packages finished [3.37s]
  1 package failed: robot_collision_checking
  1 package had stderr output: robot_collision_checking
```

---

This issue is part of a JOSS [review](https://github.com/openjournals/joss-reviews/issues/7473).
