### 2024-11-26 philip-long/ros_collision_checking#4 @mazrk7
Hi @cadojo, thanks for sharing your repository with the Docker container! I believe this has to do with the order in which Octomap and FCL are installed, where `OcTree` collision geometries are not built if Octomap is not available:
https://github.com/flexible-collision-library/fcl/blob/master/include/fcl/geometry/octree/octree.h#L43

I've updated the `README.md` to warn of this required build order.

You can update your container to install `liboctomap-dev` prior to FCL and see if that allows you to build the package. Let me know if that fixes things and I'll close the issue afterwards. Thanks!
