### 2024-11-29 philip-long/ros_collision_checking#4 @jdc-pub
This did indeed solve the build error! Thanks for the quick reply & README fix. I'll continue to tick through my review checklist. For reference, [this commit](https://github.com/cadojo/robot_collision_checking_demo/commit/e942a2d3f65974666a9e82bf0c5e293241130dd8) in my demo repository is what solved the issue: `sudo apt install liboctomap-dev` before building FCL.
