### 2024-11-29 philip-long/robot_collision_checking#5 @cadojo
Documentation Comments
- [x] What versions of ROS do you support? And what operating systems (if those operating systems are different from those supported by ROS)? Putting this information in the documentation would be helpful.
- [x] Do you have a documentation website available, ideally with an API reference? Alternatively, Markdown files with usage examples and core functions / features would be helpful.
    - I have recently found [`rosdoc2`](https://github.com/frankaemika/franka_ros2) to be a useful tool for documenting ROS packages. It's a thin wrapper around Sphinx, but you can use Sphinx themes and Markdown files to accompany the API docs.
- [x] I was happy to find the detailed [example](/src/examples/fcl_interface_example.cpp) usage with collision checking for primitives. Real-world context, e.g. solving an analysis problem, might be a really helpful addition here. I am not sure what the best real-world analysis example is for this package. ~~Maybe a static scene of a lunar landscape (spheres for boulders), and some collision checks showing a straight-line path is safe? Or a navigating around a small crowd of people (rectangles)? Any analysis example would be helpful for users.~~
    * **Edit:** I now see in the paper that "Within the constrained_manipulability package there are more examples of using the robot_collision_checking library with URDF files"; it would be helpful to link to those examples in this package's documentation.
- [x] This is not required, but I personally view code blocks as being easier to read than numbered code lines in the [installation instructions](https://github.com/philip-long/robot_collision_checking?tab=readme-ov-file#installation).
- [x] Also not required; have you considered registering `robot_collision_checking` in the ROS index?
- [x] Some projects include Dev Container configuration files and/or Docker files to help users test out examples. I think adding a `.devcontainer/` directory could be a helpful addition for users, but this is not blocking for review.
- [x] Please add some statement about community guidelines, either in the README or elsewhere in documentation. This can be something as simple as whether or not your project welcomes contributions, how users should report issues, or what kind of support users can expect.
- [x] In the [example](/src/examples/fcl_interface_example.cpp), what is the expected output? I get the following, and it's not clear to me if this is correct (though I'd guess it is).
<summary>Example Output</summary>

```
robot_collision_checking on main
⬢ [Docker] ❯ ros2 run robot_collision_checking fcl_interface_example
[INFO] [1732900217.675708634] [fcl_interface_example]: FCL interface example to visualize collision world objects
[INFO] [1732900217.676112717] [robot_collision_checking.fcl_interface_collision_world]: Creating an FCL interface collision world
[INFO] [1732900217.677394342] [fcl_interface_example]: Successful creation of FCL collision world? Yes!
[INFO] [1732900217.677687801] [fcl_interface_example]: CYLINDER with ID 1 in collision with object with ID 3
[INFO] [1732900217.677799259] [fcl_interface_example]: MESH with ID 3 in collision with object with ID 1
[INFO] [1732900218.679762218] [fcl_interface_example]: CYLINDER with ID 1 in collision with object with ID 3
[INFO] [1732900218.680336676] [fcl_interface_example]: MESH with ID 3 in collision with object with ID 1
[INFO] [1732900219.679073510] [fcl_interface_example]: CYLINDER with ID 1 in collision with object with ID 3
[INFO] [1732900219.679460302] [fcl_interface_example]: MESH with ID 3 in collision with object with ID 1
[INFO] [1732900220.678923010] [fcl_interface_example]: CYLINDER with ID 1 in collision with object with ID 3
[INFO] [1732900220.679447719] [fcl_interface_example]: MESH with ID 3 in collision with object with ID 1
[INFO] [1732900221.681658386] [fcl_interface_example]: CYLINDER with ID 1 in collision with object with ID 3
[INFO] [1732900221.682375719] [fcl_interface_example]: MESH with ID 3 in collision with object with ID 1
[INFO] [1732900222.682537303] [fcl_interface_example]: CYLINDER with ID 1 in collision with object with ID 3
[INFO] [1732900222.683021220] [fcl_interface_example]: MESH with ID 3 in collision with object with ID 1
```

---

This issue is part of a JOSS [review](https://github.com/openjournals/joss-reviews/issues/7473).
