### 2024-11-29 philip-long/robot_collision_checking#6 @cadojo
Paper Comments
- [x] Summary: Adding a sentence describing FCL to non-specialist readers would be helpful.
- [x] Summary: Adding a sentence motivating collision checking in the field of robotics might be helpful for non-specialist readers.
- [x] L21: `used by the constrained_manipulability {++ package ++}, developed...`
- [x] L22: It would be nice to link to the `constrained_manipulability` package, or provide where users can find it, if it is publicly available. If it is not publicly available, please state that too.
- [x] Statement of Need: I agree that this package provides a lower-level interface that other software suites do not. It might be worth describing how users can integrate their robots and scenes with this package. Is that functionality intended to be provided by `constrained_manipulability`? Or is that functionality "lost" in this lower-level, and is not intended to be part of this package's use-case?

---

This issue is part of a JOSS [review](https://github.com/openjournals/joss-reviews/issues/7473).
