### 2024-12-14 philip-long/ros_collision_checking#5 @mazrk7
Thank you for the detailed feedback. I have updated the README to include:
- Better information about the ROS versions/operating systems supported
- An "Examples" subsection with more on the provided example node and a link to the constrained_manipulability package
- Code blocks over numbered code lines
- A brief statement about community guidelines.

The repository also now hosts a [C++ API markdown file](https://github.com/philip-long/robot_collision_checking/blob/humble/docs/api.md) and a [Dockerfile](https://github.com/philip-long/robot_collision_checking/blob/humble/Dockerfile).

Regarding registering `robot_collision_checking` to the ROS index, we have not yet considered doing this but can do on future iterations of the project.

Thanks!
