### 2024-12-14 philip-long/ros_collision_checking#6 @mazrk7
Thank you for your comments! We have uploaded an updated version of the paper that hopefully addresses these points.
