### 2024-12-19 philip-long/ros_collision_checking#5 @jdc-pub
The updates look great! I have a couple additional minor items.

- [x] I would recommend adding some instructions relating to the Dockerfile you added. For example, a section in the README that says you can explore the package by downloading the Dockerfile, and running `docker build --tag 'robot_collision_checking' . && docker run -it 'robot_collision_checking' bash`.
- [x] Would you be able to add the Dockerfile to the other ROS 2 branch: `foxy`?
- [x] A note in the README, in the example section, for installing `rviz2` on Debian systems would also be helpful: `sudo apt install ros-$ROS_DISTRO-rviz2`.
