### 2024-12-20 philip-long/ros_collision_checking#5 @mazrk7
Great, thank you! I've added those minor recommendations to the repository:
- Instructions for the new `Dockerfile` and installation of `rviz2`
- A `foxy` branch `Dockerfile` (although please note that this is an EOL distro now)
