### 2024-12-21 philip-long/robot_collision_checking#7 @Aravind-Sundararajan
Paper Comments
- [x] The manuscript is well-written but could benefit from tightening the language to reduce repetition (e.g., repeated descriptions of FCL's capabilities across sections).
- [x] Highlight the novelty and practical impact of the package earlier in the text.
- [x] Clarify whether the target audience is primarily robotics researchers or practitioners, as this would influence the level of detail in the explanations (e.g., ROS/MoveIt integration).
- [x] Consider explicitly stating how `robot_collision_checking` simplifies tasks compared to using FCL directly.
- [x] The statement "We include ROS 1 and ROS 2 implementations" could benefit from clarifying how the two implementations differ or whether they share a common codebase.
- [x] There is no discussion of the docker images or installing ROS yourself  for currently available releases of `robot_collision_checking`
- [x] Expand briefly on the visualization capabilities. Are the visualizations limited to debugging purposes, or are they intended for runtime use?
- [x] The description of `robot_collision_checking` being used in the `constrained_manipulability` package is mentioned twice. Consolidate into one section for conciseness.
- [x] The advantages of `robot_collision_checking` over MoveIt and Pinocchio could be emphasized further. Consider a table or bullet points for clarity.
- [x] The comparison with `python-fcl` could discuss performance trade-offs between Python and C++ for collision checking.
- [x] The discussion on obstacle avoidance and path planning is helpful but would benefit from more specific, real-world use cases where robot_collision_checking excels.
- [x] Consider explicitly stating there are no conflicts of interest related to the affiliations (if applicable).

Consider including a brief section discussing potential future directions, such as:
- [x] Extending the library to support additional collision-checking algorithms or geometries.
- [x]  Supporting bindings for other languages (e.g., Python).
- [x]  Examples of integration with new robotics platforms or sensors.

This issue is part of a JOSS [review](https://github.com/openjournals/joss-reviews/issues/7473).
