### 2024-12-21 philip-long/robot_collision_checking#8 @Aravind-Sundararajan
Documentation & README Comments
README

- [x] add a link or guide for installing ROS and the more basal dependencies like eigen
- [x] alternatively, or including, add a link or instructions themselves for installing docker as depending on your setup, you mau need to setup and install from docker's apt repository
- [x] While the README assumes intermediate to advanced familiarity with ROS and robotics libraries, a brief introduction to ROS workspaces or FCL would make it more accessible to beginners.
- [x] consider appending 'For ROS' to the title so it is immediately clear that `robot_collision_checking` is a library for ROS.
- [x] Include inline links for key dependencies (e.g., ROS, RViz, URDF) to help new users quickly access them.
- [x] Clarify the unique value of this package early in the introduction. For example, emphasize its lightweight nature compared to MoveIt or its versatility with ROS 1 and ROS 2.
- [x] The instructions for choosing the correct branch could be more explicit. Add a direct command for cloning the appropriate branch: `git clone --branch <branch-name> https://github.com/<repo>`
- [x] Warn users that older versions of FCL from their ROS distribution might conflict with the required version. This is mentioned later but could be emphasized earlier.
- [x] The example is clear but could benefit from a short code snippet in the README itself to provide users with a quick preview of functionality. i.e show how a collision object is added to the world or how collision checking is invoked.
- [x] Consider adding a separate example node for dynamic scenarios.
- [x] Add a sample screenshot or gif of the RViz visualization to give users an idea of what to expect.
- [x] explan what `colcon test` and `colcon test-result` commands do and specify where test results are serialized.
- [x] briefly list or explain the types of functionality that are covered by the tests e.g. what are they and what do they do?
- [x] explain what the docker container is preconfigured with (FCL, libccd, ROS, etc) or if the user needs to do configuration.
- [x] add [contributing](https://www.contributor-covenant.org/version/2/1/code_of_conduct/code_of_conduct.md) and code of conduct https://bttger.github.io/contributing-gen-web/

- [x] possibly include a small table comparing this package to alternatives like MoveIt and Pinocchio to highlight the unique benefits (e.g., lightweight, flexible, ROS 1/2 support).
- [x] doublecheck commands and verify that backticks are used for them like `rosdep install`

I will add another issue for my notes on the API documentation

This issue is part of a JOSS [review](https://github.com/openjournals/joss-reviews/issues/7473).
