### 2024-12-21 philip-long/ros_collision_checking#5 @jdc-pub
Thank you! I was hoping to also verify seeing the collision objects in `rviz2`, but I'm failing to properly set up X11 in Docker, and I don't want that to hold up this review. I feel all of the comments here have been addressed. If you agree, please feel free to close the issue.
