### 2024-12-21 philip-long/ros_collision_checking#6 @jdc-pub
This all looks great! I think that all comments have been addressed. If you agree, please feel free to close this issue.
