### 2024-12-21 philip-long/robot_collision_checking#9 @Aravind-Sundararajan
API Documentation Comments
- [x] Some code snippets are missing syntax highlighting or have inconsistent indentation. i.e. the box example has inconsistent spacing between = and values.
- [x] `FLCInterfaceCollisionWorld code snippet could use consistent comments and formatting
- [x] Add usage examples for every major class or function. While some classes, like `FCLInterfaceCollisionWorld`, include examples in their descriptions, others, like `FCLInterfaceCollisionObject`, do not. Providing at least one usage example per key type or function will make the API more approachable.
- [x] Some assumptions are not immediately obvious. These assumptions should be stated explicitly near the relevant sections. e.g. "objects are in the same reference frame.
- [x] You reference test files like interface_test.cpp but don’t describe their contents. Add brief notes about what each test demonstrates (e.g., “tests collision between a box and sphere in the same frame”).
- [x] add some comment on error handling, state how errors are reported across the API. i.e.  what happens if `nullptr` is passed to `addCollisionObject`

This issue is part of a JOSS [review](https://github.com/openjournals/joss-reviews/issues/7473).
