### 2024-12-31 philip-long/ros_collision_checking#8 @mazrk7
Thank you for the detailed feedback @Aravind-Sundararajan.

Please see the changes from commit c424075, which addresses these points.
