### 2024-12-31 philip-long/ros_collision_checking#9 @mazrk7
Hi @Aravind-Sundararajan, can you be more specific about "consistent" or "inconsistent" comments/formatting? For example, I cannot see how the box example has inconsistent spacing. Please provide complete code snippets with the inconsistencies you're referring to, thank you.
