### 2024-12-31 philip-long/ros_collision_checking#9 @Aravind-Sundararajan
@mazrk7

For the first question, I thought on my initial review that there were inconsistent spaces and that there were missing syntax highlighting; however upon a second review I can see these review comments are erroneous or I can't see the error.  Feel free to mark this as completed.

Second, I noticed that the  example for `FCLInterfaceCollisionWorld` is the only section in the api documentation where you have inline comments, you may want to change it so these are not inline.
```
std::string frame_;  // Collision world reference frame
std::vector<FCLInterfaceCollisionObjectPtr> fcl_collision_objects_;  // List of collision objects in the world
unsigned int obj_counter_;  // Counter for object IDs
```
