### 2025-01-01 philip-long/ros_collision_checking#9 @mazrk7
@Aravind-Sundararajan great thank you for clarifying! The latest commit hopefully addresses all your points of feedback to the API documentation. Please let me know if additional edits are necessary.
