### 2025-01-01 philip-long/ros_collision_checking#9 @Aravind-Sundararajan
> @Aravind-Sundararajan great thank you for clarifying! The latest commit hopefully addresses all your points of feedback to the API documentation. Please let me know if additional edits are necessary.

Thanks looks good. I have no further edits.
