### 2025-01-01 philip-long/ros_collision_checking#9 @mazrk7
Great! I'll close this issue then.
