### 2025-01-09 philip-long/ros_collision_checking#8 @Aravind-Sundararajan
I think everything looks good now @mazrk7 !
