### 2025-01-10 philip-long/ros_collision_checking#8 @mazrk7
Great, thank you @Aravind-Sundararajan - I'll close this issue you then.
