### 2025-01-18 philip-long/ros_collision_checking#7 @mazrk7
Hi @Aravind-Sundararajan, the paper has been updated based on your suggestions. Thank you!
