### 2025-01-19 philip-long/ros_collision_checking#7 @Aravind-Sundararajan
@mazrk7 I think it looks good, I have no further comments.
