### 2025-01-19 philip-long/ros_collision_checking#7 @mazrk7
Great! Thanks @Aravind-Sundararajan, I'll close this issue now then.
