### 2025-01-23 #7473 @crvernon
👋  @mazrk7 - we are almost there! The paper is in great shape.

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the paper exactly as well). You may also add the authors' ORCID.

- [ ]  Please respond with the DOI of the archived version and the version number released here.
