### 2025-01-25 #7473 @mazrk7
Hi @crvernon, great thank you!

The DOI of the archived version on Zenodo is: [10.5281/zenodo.14736597](https://doi.org/10.5281/zenodo.14736597)

And the version number released is: **v1.0**

Please let me know if anything else is required of me.
