### 2025-01-27 #7473 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
